+++ pytsite/core/client/assets/js/common.js
@@ -1,18 +1,19 @@
 pytsite = {};
 
 pytsite.js = {
-    request: function(method, endpoint, data, success) {
+    request: function(method, endpoint, data, success, error) {
         return $.ajax({
             url: '/core/js/' + endpoint,
             method: method,
             data: data,
-            success: success
+            success: success,
+            error: error
         });
     },
-    get: function(endpoint, data, success) {
-        return pytsite.js.request('GET', endpoint, data, success)
+    get: function(endpoint, data, success, error) {
+        return pytsite.js.request('GET', endpoint, data, success, error)
     },
-    post: function(endpoint, data, success) {
-        return pytsite.js.request('POST', endpoint, data, success)
+    post: function(endpoint, data, success, error) {
+        return pytsite.js.request('POST', endpoint, data, success, error)
     }
 };